refactor: refatorando código relacionado à carrinho e sessão
