+++ src/main/java/com/esboco_comix/controller/impl/PedidoController.java
@@ -3,8 +3,11 @@ package com.esboco_comix.controller.impl;
 import com.esboco_comix.controller.utils.AbstractController;
 import com.esboco_comix.controller.utils.Router;
 import com.esboco_comix.dto.AtualizarPedidoDTO;
+import com.esboco_comix.model.Carrinho;
 import com.esboco_comix.model.entidades.Pedido;
 import com.esboco_comix.service.impl.pedido.PedidoService;
+import com.esboco_comix.sessao.SessaoService;
+
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
@@ -14,6 +17,7 @@ import java.util.Map;
 public class PedidoController extends AbstractController {
 
     private final PedidoService pedidoService = new PedidoService();
+    private final SessaoService sessaoService = new SessaoService();
 
     private final Router rotasGet = new Router(
         Map.of(
@@ -81,7 +85,8 @@ public class PedidoController extends AbstractController {
 
     private Object adicionar(HttpServletRequest req) throws Exception {
         Pedido pedido = jsonToObject(req, Pedido.class);   
-        return pedidoService.inserir(pedido, req.getSession());
+        Carrinho carrinho = sessaoService.retornarCarrinho(req.getSession());
+        return pedidoService.inserir(pedido, carrinho);
     }
 
     private Object atualizarStatus(HttpServletRequest req) throws Exception {
