--- src/main/java/com/esboco_comix/controller/session/AtributoSecao.java
@@ -1,6 +1,0 @@
-package com.esboco_comix.controller.session;
-
-public class AtributoSecao {
-    public static final String CARRINHO = "carrinho";
-}
-
